pyproject.toml
[project]
name = "voice-to-voice-agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.5",
    "fastmcp>=2.13.3",
    "langgraph>=1.0.4",
    "langgraph-cli>=0.4.7",
    "openai>=1.0.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=5.1.2",
    "tavily>=1.1.0",
    "tqdm>=4.67.1",
]




